Fix installation problem on FreeBSD 11
On FreeBSD the PREINSTALL script attempts to create Bacula user and add it to to
group with GID of 910. This works perfectly fine if the group is generated by
the bacula-client package itself, as it uses GID 910, but not when the group was
created by Puppet in common.pp, as the GID will not match. So, on FreeBSD,
ensure that GID for "bacula" group is 910.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>